requirements.txt
anyio==3.6.2
async-timeout==4.0.2
certifi==2022.12.7
fastapi==0.88.0
flake8==6.0.0
flake8-quotes==3.3.2
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.2
idna==3.4
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
numpy==1.24.1
pycodestyle==2.10.0
pydantic==1.10.4
pyflakes==3.0.1
python-telegram-bot==20.0
redis==4.4.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.45
sqlalchemy-stubs==0.4
starlette==0.22.0
types-requests==2.28.11.7
types-urllib3==1.26.25.4
typing_extensions==4.4.0
